minijules/requirements.txt
pyautogen
chromadb
sentence-transformers
tree-sitter
tree-sitter-language-pack
GitPython